Improve performance and predicate failure reporting of collateral inputs
It is possible to add a collateral input in Babbage that has a MultiAsset,
but only if there is a corresponding return TxOut that sends removes
that MultiAsset from the set of collaterals that will be collected by
the block producer. This commit makes this more apparent.

This commit also fixes the fact that when return collateral would contain
a MultiAsset then it would be reported as negative amount in the balance
in the `CollateralContainsNonADA` predicate failure.

Another part of this commit is that it improves performance of
collateral validation, because operating directly on Coin instead of Value
is more efficent because